REFACTOR: AnswerPage -> RecordAnswerPage 변경 및 ui 수정
